Default Faces For 64x32 Led Matrix
These are some default faces you can use. if you want to add your own just copy the .py file of one of them put the gif you want in the same folder make sure there are 2 one for left and one for right and you just replace the names of the gifs with yours in the .py file.