Controller script
if you have a ps4/ps5 controller or a xbox controller you can connect it to the py with bluetooth and change the faces. to add your own faces to this just change out the "blank.py" on any button with your .py file that displays your face. this does include support for a touch sensor for booping not sure if it will work for everyone so don't get your hopes up with it.