Prototype Code For Local Website For Face Control
This Has Some Prototype Code That Runs A Locally Hosted Website On The Pi Using A Hotspot(you have to setup the hotspot on your own with the pi) and it can load faces using buttons on website.
Not Sure if code works for everyone this is not a fully working version of this code so expect bugs.